debundle: split :analysis into per-subsystem crates
Split the ~29-src :analysis god-crate into three compiler-enforced
crates along the prod-dependency seams:

- :analysis (core) — ids, analysis_hints, facts/, purity/, graph,
  partition, atomic_units, factor_assembly, reports/ (schema + key
  codecs). This is one forced cluster: facts <-> purity,
  graph <-> partition, and facts -> reports::schema::SourceLocation ->
  purity -> facts / reports::schema -> factor_assembly ->
  atomic_units -> facts cycles keep them in one crate.
- :gate — realizability, validation, esm_import_order,
  rollback_graph, chunk_analysis, chunk_factorization, plus the
  ChunkFactorization-driven report builders (report_builders.rs,
  split out of reports/mod.rs which otherwise forced a core->gate
  edge) and the analysis_tests integration test module.
- :stage_one — stage_one/ composer + chunk_admission (moved into
  stage_one/, its only prod consumer).

Consumers (lowering, cli, peel, main, peel_quotient_integration_test)
now import realizability/validation/factorization symbols from gate
and Stage A symbols from stage_one directly. Pure move: no behavior
change.

BAZEL_TEST_INVOCATIONS=buildbuddy:c4db5d2e-0631-4e41-aa33-e60d64df771e
