+++ devinfra/js/debundle/analysis_tests/mod.rs
@@ -1,6 +1,6 @@
 use std::collections::{BTreeMap, BTreeSet};
 
-use crate::*;
+use analysis::*;
 use swc_common::{FileName, SyntaxContext, sync::Lrc};
 use swc_ecma_ast::*;
 use swc_ecma_parser::{Parser, StringInput, Syntax, lexer::Lexer};
