+++ devinfra/js/debundle/factor_assembly.rs
@@ -34,7 +34,7 @@ pub struct ConflictingClaim {
 ///
 /// Atom-level (binding members claimed across distinct destinations),
 /// not SCC-level. The module-quotient-SCC counterpart is
-/// [`crate::realizability::SccDiagnosis`]; the two share the
+/// `SccDiagnosis` (gate crate); the two share the
 /// "unrealizable, here's why" framing but operate on different
 /// domains (atomic unit membership vs. inter-module constraining
 /// cycles) and intentionally remain separate types.
