+++ devinfra/js/debundle/reports/schema.rs
@@ -190,7 +190,7 @@ pub struct QuotientEdgeReport {
 
 /// Wire-format projection of one quotient SCC. The in-memory
 /// primitive for **unrealizable** SCCs is
-/// [`crate::realizability::SccDiagnosis`]; this shape covers **every**
+/// `SccDiagnosis` (gate crate); this shape covers **every**
 /// SCC the dep graph turns up (single-module non-self-loops are
 /// filtered out at the builder), with `realizable` distinguishing the
 /// offending ones, and uses wire-stable string ids (`module_key`,
